Different way of reading data from files
