+++ FileReader.py
@@ -4,3 +4,9 @@ class FileReader:
         lines = file.readlines()
         file.close()
         return lines
+
+    def ReadOneLine(self, fileName):
+        file = open(fileName, 'r')
+        line = file.readline()
+        file.close()
+        return line
