+++ MyMailSender.py
@@ -3,11 +3,9 @@ from email.mime.multipart import MIMEMultipart
 from email.mime.text import MIMEText
 from FileReader import FileReader
 from MessagesFormer import MessagesFormer
+from LoginData import LoginData
 
 class MyMailSender:
-    def __init__(self):
-        self.SetLoginDataByFileName()
-
     def Send(self, messages):
         for message in messages:
             text = message.as_string()
@@ -15,15 +13,11 @@ class MyMailSender:
             messTo = message['To']
             messFrom = message['From']
 
+            loginData = LoginData()
+
             session = smtplib.SMTP_SSL('smtp.gmail.com', 465)
-            session.login(self.login, self.password)
+            session.login(loginData.GetLogin(), loginData.GetPassword())
             session.sendmail(messFrom, messTo, text)
 
         session.quit()
 
-    def SetLoginDataByFileName(self):
-        fileReader = FileReader()
-        loginData = fileReader.ReadLines("loginData.txt")
-        self.login = loginData[0]
-        self.password = loginData[1]
-
