+++ senderAddress.txt
@@ -1,0 +1,1 @@
+krzysztof.tomkow@gmail.com
